Improved code: added LocationManger.kt, DataBinding.
