+++ app/src/main/java/com/blogspot/soyamr/weathermap/presentation/map/MapsFragment.kt
@@ -1,88 +1,74 @@
 package com.blogspot.soyamr.weathermap.presentation.map
 
-import android.Manifest
 import android.app.Activity.RESULT_OK
-import android.content.IntentSender
-import android.content.pm.PackageManager
 import android.location.Location
 import android.os.Bundle
-import android.os.Looper
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.Toast
+import androidx.activity.result.ActivityResultLauncher
 import androidx.activity.result.IntentSenderRequest
 import androidx.activity.result.contract.ActivityResultContracts
-import androidx.core.app.ActivityCompat
+import androidx.core.view.isVisible
 import androidx.fragment.app.Fragment
 import com.blogspot.soyamr.weathermap.R
-import com.google.android.gms.common.api.ResolvableApiException
-import com.google.android.gms.location.*
+import com.blogspot.soyamr.weathermap.databinding.FragmentMapsBinding
 import com.google.android.gms.maps.CameraUpdateFactory
 import com.google.android.gms.maps.GoogleMap
 import com.google.android.gms.maps.OnMapReadyCallback
 import com.google.android.gms.maps.SupportMapFragment
 import com.google.android.gms.maps.model.LatLng
 import com.google.android.gms.maps.model.MarkerOptions
-import com.google.android.gms.tasks.Task
 
 class MapsFragment : Fragment() {
 
-    private lateinit var fusedLocationClient: FusedLocationProviderClient
+    private lateinit var binding: FragmentMapsBinding
+
     private lateinit var googleMap: GoogleMap
 
-    private val locationCallback: LocationCallback by lazy {
-        object : LocationCallback() {
-            override fun onLocationResult(locationResult: LocationResult?) {
-                locationResult ?: return
-                for (location in locationResult.locations) {
-                    if (location != null) {
-                        addMarker(location)
-                        fusedLocationClient.removeLocationUpdates(this)
-                        break
-                    }
-                }
-            }
-        }
+    private val locationManger: LocationManger by lazy {
+        LocationManger(
+            requireContext(),
+            addMarker,
+            resolutionForResult,
+            requestPermissionLauncher,
+            showMessage
+        )
     }
 
-    private val requestPermissionLauncher =
+    //callbacks
+    private val addMarker: (userLocation: Location) -> Unit =
+        {
+            binding.progressCircular.isVisible = false
+            val userLocation = LatLng(it.latitude, it.longitude)
+            googleMap.addMarker(MarkerOptions().position(userLocation).title("you are Here "))
+            googleMap.moveCamera(CameraUpdateFactory.newLatLng(userLocation))
+            googleMap.animateCamera(CameraUpdateFactory.zoomTo(15F));
+        }
+
+    private val requestPermissionLauncher: ActivityResultLauncher<String> =
         registerForActivityResult(ActivityResultContracts.RequestPermission()) { isGranted: Boolean ->
             if (isGranted) {
-                setMarkerOnMap()
+                locationManger.setMarkerOnMap()
             } else {
-                showMessage(R.string.can_not_find_location)
+                showMessage(R.string.can_not_find_location,false)
             }
         }
 
     private val callback = OnMapReadyCallback { googleMap ->
         this.googleMap = googleMap
-        setMarkerOnMap()
+        locationManger.setMarkerOnMap()
     }
 
-    private val locationRequest: LocationRequest by lazy {
-        LocationRequest.create().apply {
-            interval = 10000
-            fastestInterval = 5000
-            priority = LocationRequest.PRIORITY_HIGH_ACCURACY
-        }
-    }
-    private val resolutionForResult =
+    private val resolutionForResult: ActivityResultLauncher<IntentSenderRequest> =
         registerForActivityResult(ActivityResultContracts.StartIntentSenderForResult()) { activityResult ->
-            if (activityResult.resultCode == RESULT_OK)
-                startLocationUpdates()
-            else {
-                showMessage(R.string.can_not_find_location)
+            if (activityResult.resultCode == RESULT_OK) {
+                locationManger.startLocationUpdates()
+            } else {
+                showMessage(R.string.can_not_find_location,false)
             }
         }
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        return inflater.inflate(R.layout.fragment_maps, container, false)
-    }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
@@ -90,78 +76,25 @@ class MapsFragment : Fragment() {
         mapFragment?.getMapAsync(callback)
     }
 
-    private fun showMessage(msgId: Int) {
-        Toast.makeText(requireContext(), getString(msgId), Toast.LENGTH_SHORT).show()
-    }
-
-    private fun setMarkerOnMap() {
-        if (ActivityCompat.checkSelfPermission(
-                requireContext(),
-                Manifest.permission.ACCESS_FINE_LOCATION
-            )
-            != PackageManager.PERMISSION_GRANTED
-        ) {
-            requestPermissionLauncher.launch(Manifest.permission.ACCESS_FINE_LOCATION)
-            return
+    val showMessage: (msgId: Int, progressBarVisible: Boolean) -> Unit =
+        { mgsId, isVisible ->
+            binding.progressCircular.isVisible = isVisible
+            Toast.makeText(requireContext(), getString(mgsId), Toast.LENGTH_SHORT).show()
         }
-        fusedLocationClient = LocationServices.getFusedLocationProviderClient(requireContext())
-        fusedLocationClient.lastLocation.addOnSuccessListener {
-            if (it != null) {
-                addMarker(it)
-            } else {
-                requestLocationUpdates()
-            }
-        }
-    }
-
-    private fun requestLocationUpdates() {
-        val builder = LocationSettingsRequest.Builder()
-            .addLocationRequest(locationRequest)
-        val client: SettingsClient = LocationServices.getSettingsClient(requireContext())
-        val task: Task<LocationSettingsResponse> = client.checkLocationSettings(builder.build())
-
-        task.addOnSuccessListener { locationSettingsResponse ->
-            startLocationUpdates()
-        }
-
-        task.addOnFailureListener { exception ->
-            if (exception is ResolvableApiException) {
-                try {
-                    val intentSenderRequest =
-                        IntentSenderRequest.Builder(exception.resolution).build()
-                    resolutionForResult.launch(intentSenderRequest)
-
-                } catch (sendEx: IntentSender.SendIntentException) {
-                    // Ignore the error.
-                }
-            }
-        }
-    }
-
-    private fun addMarker(userLocation: Location) {
-        val userLocationLatLng = LatLng(userLocation.latitude, userLocation.longitude)
-        googleMap.addMarker(MarkerOptions().position(userLocationLatLng).title("you are Here "))
-        googleMap.moveCamera(CameraUpdateFactory.newLatLng(userLocationLatLng))
-        googleMap.animateCamera(CameraUpdateFactory.zoomTo(15F));
-    }
-
-    private fun startLocationUpdates() {
-        if (ActivityCompat.checkSelfPermission(
-                requireContext(),
-                Manifest.permission.ACCESS_FINE_LOCATION
-            ) != PackageManager.PERMISSION_GRANTED
-        ) {
-            return
-        }
-        fusedLocationClient.requestLocationUpdates(
-            locationRequest,
-            locationCallback,
-            Looper.getMainLooper()
-        )
-    }
 
     override fun onPause() {
         super.onPause()
-        fusedLocationClient.removeLocationUpdates(locationCallback)
+        locationManger.stopLocationUpdates()
+    }
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        return FragmentMapsBinding.inflate(inflater).run {
+            binding = this
+            root
+        }
     }
 }
