+++ app/src/main/res/values/strings.xml
@@ -4,4 +4,5 @@
     <string name="hello_blank_fragment">Hello blank fragment</string>
     <string name="no_internet">Sorry you are not connected to internet:)</string>
     <string name="can_not_find_location">we can\'t get your location on the map</string>
+    <string name="getting_your_location">getting your current location wait..</string>
 </resources>
